requirements.txt
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
cycler==0.11.0
Cython==0.29.26
deepl==1.3.1
ekphrasis==0.5.1
filelock==3.4.0
fonttools==4.28.5
ftfy==6.0.3
hdbscan==0.8.27
huggingface-hub==0.2.1
idna==3.3
importlib-resources==5.4.0
joblib==1.1.0
JsonForm==0.0.2
jsonschema==4.4.0
JsonSir==0.0.2
kiwisolver==1.3.2
llvmlite==0.36.0
matplotlib==3.5.1
nltk==3.6.5
numba==0.53.1
numpy==1.22.0
oauthlib==3.1.1
packaging==21.3
pandas==1.3.5
Pillow==8.4.0
pyarrow==6.0.1
pynndescent==0.5.5
pyparsing==3.0.6
pyreadr==0.4.4
pyrsistent==0.18.1
python-dateutil==2.8.2
Python-EasyConfig==0.1.7
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
Resource==0.2.1
sacremoses==0.0.46
scikit-learn==1.0.1
scipy==1.7.3
sentence-transformers==2.1.0
sentencepiece==0.1.96
six==1.16.0
termcolor==1.1.0
threadpoolctl==3.0.0
tokenizers==0.10.3
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
tqdm==4.62.3
transformers==4.14.1
tweepy==4.4.0
tweet-preprocessor==0.6.0
typing_extensions==4.0.1
ujson==5.0.0
umap-learn==0.5.2
urllib3==1.26.7
wcwidth==0.2.5
zipp==3.7.0
